move spec_template to its own module
